init：使用mybatis完成新增,查询

+++ src/main/java/com/sinnevy/minproject/vo/PageVo.java
@@ -12,17 +12,16 @@ public class PageVo<T> {
 
     private Integer pageSize;
 
-    private Integer total;
+    private Long total;
 
     private List<T> list;
 
-    public PageVo<T> init(Integer page, Integer pageSize, Integer total) {
+    public PageVo<T> init(Integer page, Integer pageSize, Long total) {
         this.page = page;
         this.pageSize = pageSize;
         this.total = total;
         this.list = new ArrayList<>();
         return this;
-
     }
